+++ src/matcher_combinators/matchers.clj
@@ -7,16 +7,16 @@
   [expected]
   (cond
     (sequential? expected) (core/->EqualsSeq expected)
-    (set? expected)        (core/->EqualsSet expected false)
+    (set? expected)        (core/->SetEquals expected false)
     (map? expected)        (core/->EqualsMap expected)
     :else                  (core/->Value expected)))
 
-(defn equals-set
+(defn set-equals
   "Matches a set in the way `(equals some-set)` would, but accepts sequences as
   the expected matcher argument, allowing one to use matchers with the same
   submatcher appearing more than once."
   [expected]
-  (core/->EqualsSet expected true))
+  (core/->SetEquals expected true))
 
 (defn embeds
   "Matcher that will match when the map contains some of the same key/values as
@@ -24,16 +24,16 @@
   [expected]
   (cond
     (sequential? expected) (core/->EmbedsSeq expected)
-    (set? expected)        (core/->EmbedsSet expected false)
+    (set? expected)        (core/->SetEmbeds expected false)
     (map? expected)        (core/->EmbedsMap expected)
     :else                  (core/->InvalidType expected "embeds" "seq, set, map")))
 
-(defn embeds-set
+(defn set-embeds
   "Matches a set in the way `(embeds some-set)` would, but accepts sequences
   as the expected matcher argument, allowing one to use matchers with the same
   submatcher appearing more than once."
   [expected]
-  (core/->EmbedsSet expected true))
+  (core/->SetEmbeds expected true))
 
 (defn in-any-order
   "Matcher that will match when the given a list that is the same as the
@@ -48,10 +48,10 @@
   ([select-fn expected]
    (core/->SelectingInAnyOrder select-fn expected)))
 
-(defn prefix-seq
+(defn prefix
   "Matcher that will match when provided a (ordered) prefix of the `expected`
   list.
 
   Similar to Midje's `(embeds expected)`"
   [expected]
-  (core/->PrefixSeq expected))
+  (core/->Prefix expected))